requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
datasets==4.2.0
deepspeed==0.18.0
dill==0.4.0
distro==1.9.0
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
h11==0.16.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
jinja2==3.1.6
jiter==0.11.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
markupsafe==3.0.3
mathruler==0.1.0
matplotlib==3.10.7
mdurl==0.1.2
mpi4py==4.1.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.127
openai==2.6.1
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.12.2
pydantic-core==2.41.4
pygments==2.19.2
pylatexenc==2.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
qwen-vl-utils==0.0.14
regex==2025.9.18
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
scipy==1.15.3
sentry-sdk==2.43.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.4
tqdm==4.67.1
triton==3.2.0
trl==0.15.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
wandb==0.22.3
xxhash==3.6.0
yarl==1.22.0

# 单独装torch和flash-attn
# pip install -r requirements.txt
# pip install torch==2.6.0 torchvision==0.21.0 torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu124
# pip install flash-attn==2.7.4.post1 --no-build-isolation